dug/requirements.txt
numpy==1.24.3
pandas==2.0.2
python-dateutil==2.8.2
streamlit==1.22.0

black==23.3.0
isort==5.12.0
pandas-stubs==2.0.1.230501
playwright==1.34.0
pylint==2.17.4
pyright==1.1.310
pytest-playwright==0.3.3
pytest==7.3.1
types-python-dateutil==2.8.19.13
